+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Dependency
 
-- TexTransCore が v0.2.x を要求する用意なりました。 (#1016)
+- TexTransCore v0.2.x を要求する用意なりました。 (#1016)
 
 ## [v0.10.6](https://github.com/ReinaS-64892/TexTransTool/compare/v0.10.5...v0.10.6) - 2025-05-28